file relocate and fix few typo
