+++ src/backend/.gitignore
@@ -6,3 +6,4 @@ __pycache__/
 .venv
 
 account_key.json
+sql_app.db
